requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
mysql-connector-python
alembic
python-dotenv
pytest
requests